refactor(blood_request): restructure blood request and campaign entities and services
feat(blood_request): add separate endpoints for blood requests and campaigns
fix(donor_schedules): enforce user ownership checks for donor schedules
style: remove unused code and clean up whitespace

+++ DarahConnectAPI/internal/http/handler/donor_schedules.go
@@ -3,8 +3,10 @@ package handler
 import (
 	"net/http"
 
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/labstack/echo/v4"
 	"github.com/mhusainh/DarahConnect/DarahConnectAPI/pkg/response"
+	"github.com/mhusainh/DarahConnect/DarahConnectAPI/pkg/token"
 
 	"github.com/mhusainh/DarahConnect/DarahConnectAPI/internal/http/dto"
 	"github.com/mhusainh/DarahConnect/DarahConnectAPI/internal/service"
@@ -26,7 +28,19 @@ func (h *DonorScheduleHandler) GetDonorSchedules(ctx echo.Context) error {
 		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
 	}
 
-	donorSchedules, total, err := h.donorScheduleService.GetAll(ctx.Request().Context(), req)
+	// Retrieve user claims from the JWT token
+	claims, ok := ctx.Get("user").(*jwt.Token)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user claims")
+	}
+
+	// Extract user information from claims
+	claimsData, ok := claims.Claims.(*token.JwtCustomClaims)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user information from claims")
+	}
+
+	donorSchedules, total, err := h.donorScheduleService.GetAll(ctx.Request().Context(), claimsData.Id, req)
 	if err != nil {
 		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
 	}
@@ -39,29 +53,27 @@ func (h *DonorScheduleHandler) GetDonorSchedule(ctx echo.Context) error {
 		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
 	}
 
+	// Retrieve user claims from the JWT token
+	claims, ok := ctx.Get("user").(*jwt.Token)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user claims")
+	}
+
+	// Extract user information from claims
+	claimsData, ok := claims.Claims.(*token.JwtCustomClaims)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user information from claims")
+	}
+
 	donorSchedule, err := h.donorScheduleService.GetById(ctx.Request().Context(), req.Id)
 	if err != nil {
 		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
 	}
+
+	if claimsData.Id != donorSchedule.UserId {
+		return ctx.JSON(http.StatusUnauthorized, response.ErrorResponse(http.StatusUnauthorized, "unauthorized"))
+	}
 	return ctx.JSON(http.StatusOK, response.SuccessResponse("successfully showing donor schedule", donorSchedule))
-}
-
-func (h *DonorScheduleHandler) GetDonorScheduleByHospitalId(ctx echo.Context) error {
-	var req dto.GetAllDonorScheduleRequest
-	if err := ctx.Bind(&req); err != nil {
-		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
-	}
-
-	var hospitalId dto.HospitalGetByIdRequest
-	if err := ctx.Bind(&hospitalId); err != nil {
-		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
-	}
-
-	donorSchedules, total, err := h.donorScheduleService.GetByHospitalId(ctx.Request().Context(), hospitalId.Id, req)
-	if err != nil {
-		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
-	}
-	return ctx.JSON(http.StatusOK, response.SuccessResponseWithPagi("successfully showing all donor schedules", donorSchedules, req.Page, req.Limit, total))
 }
 
 func (h *DonorScheduleHandler) CreateDonorSchedule(ctx echo.Context) error {
@@ -69,6 +81,20 @@ func (h *DonorScheduleHandler) CreateDonorSchedule(ctx echo.Context) error {
 	if err := ctx.Bind(&req); err != nil {
 		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
 	}
+
+	// Retrieve user claims from the JWT token
+	claims, ok := ctx.Get("user").(*jwt.Token)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user claims")
+	}
+
+	// Extract user information from claims
+	claimsData, ok := claims.Claims.(*token.JwtCustomClaims)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user information from claims")
+	}
+
+	req.UserId = claimsData.Id
 
 	if err := h.donorScheduleService.Create(ctx.Request().Context(), req); err != nil {
 		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
@@ -82,9 +108,25 @@ func (h *DonorScheduleHandler) UpdateDonorSchedule(ctx echo.Context) error {
 		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
 	}
 
+	// Retrieve user claims from the JWT token
+	claims, ok := ctx.Get("user").(*jwt.Token)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user claims")
+	}
+
+	// Extract user information from claims
+	claimsData, ok := claims.Claims.(*token.JwtCustomClaims)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user information from claims")
+	}
+
 	donorSchedule, err := h.donorScheduleService.GetById(ctx.Request().Context(), req.Id)
 	if err != nil {
 		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
+	}
+
+	if claimsData.Id != donorSchedule.UserId {
+		return ctx.JSON(http.StatusUnauthorized, response.ErrorResponse(http.StatusUnauthorized, "unauthorized"))
 	}
 
 	if err := h.donorScheduleService.Update(ctx.Request().Context(), req, donorSchedule); err != nil {
@@ -99,6 +141,27 @@ func (h *DonorScheduleHandler) DeleteDonorSchedule(ctx echo.Context) error {
 		return ctx.JSON(http.StatusBadRequest, response.ErrorResponse(http.StatusBadRequest, err.Error()))
 	}
 
+	// Retrieve user claims from the JWT token
+	claims, ok := ctx.Get("user").(*jwt.Token)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user claims")
+	}
+
+	// Extract user information from claims
+	claimsData, ok := claims.Claims.(*token.JwtCustomClaims)
+	if !ok {
+		return ctx.JSON(http.StatusInternalServerError, "unable to get user information from claims")
+	}
+
+	donorSchedule, err := h.donorScheduleService.GetById(ctx.Request().Context(), req.Id)
+	if err != nil {
+		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
+	}
+
+	if claimsData.Id != donorSchedule.UserId {
+		return ctx.JSON(http.StatusUnauthorized, response.ErrorResponse(http.StatusUnauthorized, "unauthorized"))
+	}
+	
 	if err := h.donorScheduleService.Delete(ctx.Request().Context(), req.Id); err != nil {
 		return ctx.JSON(http.StatusInternalServerError, response.ErrorResponse(http.StatusInternalServerError, err.Error()))
 	}
